+++ packages/Datadog.Unity/Editor/iOS/PostBuildProcess.cs
@@ -98,7 +98,7 @@ namespace Datadog.Unity.Editor.iOS
 import Foundation
 import DatadogCore
 import DatadogLogs
-import DatadogRUM
+@_spi(Experimental) import DatadogRUM
 import DatadogCrashReporting
 
 @_cdecl(""initializeDatadog"")
@@ -181,8 +181,8 @@ func initializeDatadog() {{
 
                 if (options.EnableTimeseries)
                 {
-                    sb.AppendLine("    rumConfig.enableTimeseries = true");
-                    sb.AppendLine($"    rumConfig.timeseriesBatchSize = {options.TimeseriesBatchSize}");
+                    sb.AppendLine(
+                        $"    rumConfig.timeseries = RUM.Configuration.Timeseries(collectTypes: {GetSwiftTimeseriesCollectTypes(options.TimeseriesTypes)})");
                 }
 
                 // Uncomment to enable RUM Configuration Telemetry
@@ -347,6 +347,27 @@ find . -type d -name '*.dSYM' -exec cp -r '{{}}' ""$PROJECT_DIR/{SymbolAssemblyB
                 VitalsUpdateFrequency.Frequent => ".frequent",
                 _ => "nil",
             };
+        }
+
+        private static string GetSwiftTimeseriesCollectTypes(TimeseriesTypes collectTypes)
+        {
+            if (collectTypes == TimeseriesTypes.All)
+            {
+                return "nil";
+            }
+
+            var swiftTypes = new List<string>();
+            if (collectTypes.HasFlag(TimeseriesTypes.Memory))
+            {
+                swiftTypes.Add(".memory");
+            }
+
+            if (collectTypes.HasFlag(TimeseriesTypes.Cpu))
+            {
+                swiftTypes.Add(".cpu");
+            }
+
+            return $"[{string.Join(", ", swiftTypes)}]";
         }
 
         private static string GetSwiftSite(DatadogSite site)