+++ packages/Datadog.Unity/Tests/Editor/DatadogConfigurationOptionsTests.cs
@@ -75,7 +75,7 @@ namespace Datadog.Unity.Editor.Tests
             Assert.AreEqual(false, options.TrackNonFatalAppHangs);
             Assert.AreEqual(0.25f, options.NonFatalAppHangThreshold);
             Assert.IsFalse(options.EnableTimeseries);
-            Assert.AreEqual(120, options.TimeseriesBatchSize);
+            Assert.AreEqual(TimeseriesTypes.All, options.TimeseriesTypes);
         }
     }
 }